Separate ExceptionHandlingTest into existing test cases within an ExceptionHandling namespace
